Fix warnings related to use of obsolete colorpalette
